ref : move products type to service products type file
